+++ src/libcsg/modules/io/lammpsdumpwriter.cc
@@ -56,9 +56,9 @@ void LAMMPSDumpWriter::Write(Topology *conf) {
        iter != conf->Beads().end(); ++iter) {
     Bead *bi = *iter;
 
-    int type_id = conf->getBeadTypeId(bi->getType());
+    long int type_id = conf->getBeadTypeId(bi->getType());
 
-    fprintf(_out, "%ld %i", bi->getId() + 1, type_id);
+    fprintf(_out, "%ld %li", bi->getId() + 1, type_id);
     fprintf(_out, " %f %f %f", bi->getPos().x() * conv::nm2ang,
             bi->getPos().y() * conv::nm2ang, bi->getPos().z() * conv::nm2ang);
     if (v) {
